Moving server logger to internal
